split solution into logic and main files inside src directory
